cleanup and color changes
