+++ src/lantern/RevocationsByGender/RevocationsByGender.js
@@ -25,17 +25,17 @@ import HorizontalBarChartWithLabels from "../BarCharts/HorizontalBarChartWithLab
 import createGenerateChartData from "./createGenerateChartData";
 import { useRootStore } from "../../components/StoreProvider";
 import { genderValueToLabel } from "../../utils/formatStrings";
+import { US_PA } from "../../RootStore/TenantStore/lanternTenants";
 
 const DEFAULT_MODE = "MALE";
 
 const RevocationsByGender = observer(
   ({ containerHeight, timeDescription }, ref) => {
-    const { dataStore } = useRootStore();
-    const { currentTenantId } = useRootStore();
+    const { dataStore, currentTenantId } = useRootStore();
     const { revocationsChartStore } = dataStore;
     const CHART_TITLE = translate("revocationsByGenderChartTitle");
     const CHART_ID = translate("revocationsByGenderChartId");
-    const stacked = currentTenantId === "US_PA";
+    const stacked = currentTenantId === US_PA;
 
     return (
       <RevocationsByDimension
