+++ src/lantern/RevocationsByRace/RevocationsByRace.js
@@ -24,6 +24,7 @@ import createGenerateChartData from "./createGenerateChartData";
 import { translate } from "../../utils/i18nSettings";
 import { useDataStore, useRootStore } from "../../components/StoreProvider";
 import HorizontalBarChartWithLabels from "../BarCharts/HorizontalBarChartWithLabels";
+import { US_PA } from "../../RootStore/TenantStore/lanternTenants";
 
 const DEFAULT_MODE = "WHITE";
 
@@ -35,7 +36,7 @@ const RevocationsByRace = observer(
     const { revocationsChartStore } = dataStore;
     const CHART_TITLE = translate("revocationsByRaceChartTitle");
     const CHART_ID = translate("revocationsByRaceChartId");
-    const stacked = currentTenantId === "US_PA";
+    const stacked = currentTenantId === US_PA;
 
     return (
       <RevocationsByDimension
